Cleaned unwanted functions and marked TODOs
Removed unecessary code
