+++ SampleActivity/Sample/src/main/res/layout/activity_main.xml
@@ -57,25 +57,25 @@
             android:paddingTop="10dp"
             android:paddingBottom="10dp" />
 
-        <Button
-            android:id="@+id/stopRoute"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Stop Route"
-            android:textStyle="bold"
-            android:layout_below="@+id/spinner"
-            android:layout_alignRight="@+id/zoomin"
-            android:layout_alignEnd="@+id/zoomin" />
+        <!--<Button-->
+            <!--android:id="@+id/stopRoute"-->
+            <!--android:layout_width="wrap_content"-->
+            <!--android:layout_height="wrap_content"-->
+            <!--android:text="Stop Route"-->
+            <!--android:textStyle="bold"-->
+            <!--android:layout_below="@+id/spinner"-->
+            <!--android:layout_alignRight="@+id/zoomin"-->
+            <!--android:layout_alignEnd="@+id/zoomin" />-->
 
-        <Button
-            android:id="@+id/pinMarker"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Place a Pin Marker"
-            android:textStyle="bold"
-            android:layout_below="@+id/spinner"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentStart="true" />
+        <!--<Button-->
+            <!--android:id="@+id/pinMarker"-->
+            <!--android:layout_width="wrap_content"-->
+            <!--android:layout_height="wrap_content"-->
+            <!--android:text="Place a Pin Marker"-->
+            <!--android:textStyle="bold"-->
+            <!--android:layout_below="@+id/spinner"-->
+            <!--android:layout_alignParentLeft="true"-->
+            <!--android:layout_alignParentStart="true" />-->
 
         <TextView
             android:layout_width="wrap_content"
@@ -103,7 +103,6 @@
             android:padding="5dp"
             android:textStyle="bold"
             android:background="#80000000"
-            android:layout_below="@+id/stopRoute"
             android:layout_alignParentStart="false"
             android:layout_alignParentTop="false"
             android:layout_alignParentBottom="false" />
