+++ SampleActivity/Sample/src/main/res/menu/main.xml
@@ -4,15 +4,15 @@
         android:orderInCategory="100"
         android:showAsAction="never" />
 
-    <item android:id="@+id/start_location_engine"
-        android:title="@string/start_location_engine"
-        android:orderInCategory="100"
-        android:icon="@drawable/start_location"
-        android:showAsAction="ifRoom" />
+    <!--<item android:id="@+id/start_location_engine"-->
+        <!--android:title="@string/start_location_engine"-->
+        <!--android:orderInCategory="100"-->
+        <!--android:icon="@drawable/start_location"-->
+        <!--android:showAsAction="ifRoom" />-->
 
-    <item android:id="@+id/stop_location_engine"
-        android:title="@string/stop_location_engine"
-        android:orderInCategory="100"
-        android:icon="@drawable/stop_location"
-        android:showAsAction="ifRoom" />
+    <!--<item android:id="@+id/stop_location_engine"-->
+        <!--android:title="@string/stop_location_engine"-->
+        <!--android:orderInCategory="100"-->
+        <!--android:icon="@drawable/stop_location"-->
+        <!--android:showAsAction="ifRoom" />-->
 </menu>
